Optimize d20
It annoyed me that it took more than 4 seconds to solve p2. With these
optimizations it takes less than 2. Still not great but I cannot see
any more low hanging fruit.

The new implementation of int_from_digits takes 1/3 as long as the old.
Not that it matters much since I went on to cache it. This accounted
for about 30% of the original runtime.

Using a defaultdict instead of get shaved about 3pp of the duration of
the list comprehension. Not much but it turned out to also make the
tuple construction more concise.

Constructing a tuple was necessary in order to benefit from caching the
conversion function. However, creating a tuple from a comprehension is
much slower than creating a list. Enough so to almost offset all of the
benefit from the caching in the first place. At this point I had
already precomputed the neighbor offsets and inlined their addition,
which resulted in a big improvement. Using a fixed size tuple improved
this even more.

Now the only thing that stood out was the computation of minmax taking
about 5% of the runtime. Since eliminating it was straight forward I
did so.