Optimize d23
The biggest improvement comes from rewriting the search to allow
caching. I struggled for a bit to switch mindset and kept getting the
wrong result because I returned the cost including upstream cost.

Also prunes the tree more aggressively. Number of calls is down to
about 2M (would have been about 3M without the improved pruning) and
the time for the test suite is down to about 30s.