Some micro optimizations for d23
The biggest benefit comes from pre-computing the order in which to try
hallway locations and terminating early if one is blocked.

Optimizing the blocked function for doing at most one move helped a
little but that certainly may be overfitting on my examples. A similar
analysis could be implied to changing the order of pruning conditions.

Moving computation to after pruning should be risk free though, even if
the improvements where not really observable.

Total time for test suite is now 25s.