Use tuples instead of lists for d23
This guarantees that callees will not mutate them which makes me a
little less nervous about not doing effectively deep copies of the
rooms dict. This enables the more stylish and, surprisingly, faster
dict union to be used instead of a comprehension.

Also optimize and inline cost function as a micro optimization. It
took 2% of runtime for only a few lines, so maybe this made a
difference. I honestly did not profile this change carefully. In any
case it does not hurt readability imo so harm should be limited.

Total time for test suite is now 20s.