Separate herds in d25 for speed
When I saw the size of the input I thought it was designed to make a
naive algorithm too slow. The first thing that came to mind was
splitting the herds but when it did not work I abandoned it for the
simpler one-herd solution. Now that I have found that the problem was
a column mixed up with a row it is trivial to add it back. It does half
the runtime which is nice. It would of course not have helped if the
input truly was too big.

Duration: 1.3s